examples/01-basic-chat/requirements.txt
djangosdk>=0.1.1
whitenoise>=6.0
python-decouple>=3.8
